[TASK] Tweak schema migrations to Flow/Party decoupling
This adjusts schema migrations to that it is actually possible to:

* install Flow without Party
* add Party to a "clean" Flow setup

Releases: master, 3.0
Related: FLOW-5
Original-Commit-Hash: 33e496bba42cf3863354ab1eb11343e762016e9c

+++ TYPO3.Flow/Migrations/Mysql/Version20150206114820.php
@@ -17,10 +17,12 @@ class Version20150206114820 extends AbstractMigration {
 	 */
 	public function up(Schema $schema) {
 		$this->abortIf($this->connection->getDatabasePlatform()->getName() != "mysql");
-		$this->abortIf($this->isPartyPackageInstalled() && !$this->isCorrespondingPartyMigrationExecuted(), 'This migration requires a current version of the TYPO3.Party package and its migration 20150206113911 being applied.');
 
-		$this->addSql("ALTER TABLE typo3_flow_security_account DROP FOREIGN KEY typo3_flow_security_account_ibfk_1");
-		$this->addSql("DROP INDEX IDX_65EFB31C89954EE0 ON typo3_flow_security_account");
+		if ($this->isPartyPackageInstalled()) {
+			$this->addSql("ALTER TABLE typo3_flow_security_account DROP FOREIGN KEY typo3_flow_security_account_ibfk_1");
+			$this->addSql("DROP INDEX IDX_65EFB31C89954EE0 ON typo3_flow_security_account");
+		}
+
 		$this->addSql("ALTER TABLE typo3_flow_security_account DROP party");
 	}
 
@@ -32,8 +34,11 @@ class Version20150206114820 extends AbstractMigration {
 		$this->abortIf($this->connection->getDatabasePlatform()->getName() != "mysql");
 
 		$this->addSql("ALTER TABLE typo3_flow_security_account ADD party VARCHAR(40) DEFAULT NULL");
-		$this->addSql("ALTER TABLE typo3_flow_security_account ADD CONSTRAINT typo3_flow_security_account_ibfk_1 FOREIGN KEY (party) REFERENCES typo3_party_domain_model_abstractparty (persistence_object_identifier)");
-		$this->addSql("CREATE INDEX IDX_65EFB31C89954EE0 ON typo3_flow_security_account (party)");
+
+		if ($this->isPartyPackageInstalled()) {
+			$this->addSql("CREATE INDEX IDX_65EFB31C89954EE0 ON typo3_flow_security_account (party)");
+			$this->addSql("ALTER TABLE typo3_flow_security_account ADD CONSTRAINT typo3_flow_security_account_ibfk_1 FOREIGN KEY (party) REFERENCES typo3_party_domain_model_abstractparty (persistence_object_identifier)");
+		}
 	}
 
 	/**
@@ -41,12 +46,5 @@ class Version20150206114820 extends AbstractMigration {
 	 */
 	protected function isPartyPackageInstalled() {
 		return $this->sm->tablesExist(array('typo3_party_domain_model_abstractparty'));
-	}
-
-	/**
-	 * @return boolean
-	 */
-	protected function isCorrespondingPartyMigrationExecuted() {
-		return $this->sm->tablesExist(array('typo3_party_domain_model_abstractparty_accounts_join'));
 	}
 }
